# packages/shared-contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "stock-research-platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/research/tasks": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_v1_research_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/research/reports": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Generate Report",
        "operationId": "generate_report_api_v1_research_reports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/research/tasks/{task_id}": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_v1_research_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/tasks/{task_id}/events": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Task Events",
        "operationId": "task_events_api_v1_research_tasks__task_id__events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol"
          },
          "mode": {
            "type": "string",
            "pattern": "^(quick|standard|deep)$",
            "title": "Mode",
            "default": "standard"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "ReportRequest"
      },
      "ReportResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "module_version": {
            "type": "string",
            "title": "Module Version"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ReportSectionResponse"
            },
            "type": "array",
            "title": "Sections"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "as_of",
          "module_version",
          "summary",
          "sections"
        ],
        "title": "ReportResponse"
      },
      "ReportSectionResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "title",
          "data"
        ],
        "title": "ReportSectionResponse"
      },
...
      "TaskCreateRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol"
          },
          "mode": {
            "type": "string",
            "pattern": "^(quick|standard|deep)$",
            "title": "Mode"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modules"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "mode"
        ],
        "title": "TaskCreateRequest"
      },
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "requested_modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requested Modules"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "user_id",
          "task_type",
          "status",
          "symbol",
          "mode",
          "as_of",
          "requested_modules",
          "question",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}